% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectFrame.R
\name{Create.newBioThemeFinder.GSEA}
\alias{Create.newBioThemeFinder.GSEA}
\title{When gene names and vector measuring the chenge of specific genes, like logFC, are available, create BioThemeFinder object with FC using ORA.}
\usage{
Create.newBioThemeFinder.GSEA(
  Gene,
  log2FC,
  FromType = "SYMBOL",
  Species = "human"
)
}
\arguments{
\item{Gene}{A vector containing genes}

\item{log2FC}{A vector containing genes, the order of the above two must be the same}

\item{FromType}{Type of genes, can be the columns supported by org.DBs like SYMBOL for gene symbol or ENSEMBL for Ensembl ID}

\item{Species}{one of the human, mouse of non-gene}
}
\value{
A BioThemeFinder.GSEA object
}
\description{
When gene names and vector measuring the chenge of specific genes, like logFC, are available, create BioThemeFinder object with FC using ORA.
}
\author{
Zhiming Ye
}
