% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectFrame.R
\name{Create.newBioThemeFinder.ORAwithFC}
\alias{Create.newBioThemeFinder.ORAwithFC}
\title{When gene names and vector measuring the chenge of specific genes, like logFC, are available, create BioThemeFinder object with FC using ORA.}
\usage{
Create.newBioThemeFinder.ORAwithFC(
  Gene,
  log2FC,
  Pvalue,
  FCcutoff = 1,
  PvalueCutOff = 0.05,
  FromType = "SYMBOL",
  Species = "human",
  CntsOfDiffGene = 2
)
}
\arguments{
\item{Gene}{A vector containing genes}

\item{log2FC}{A vector containing genes}

\item{Pvalue}{A vector containing pvalue, the order of the above three must be the same}

\item{FCcutoff}{the cut off of abs(FC) value, default as 1}

\item{PvalueCutOff}{the cut off of P value, default as 0.05}

\item{FromType}{Type of genes, can be the columns supported by org.DBs like SYMBOL for gene symbol or ENSEMBL for Ensembl ID}

\item{Species}{one of the human, mouse of non-gene}

\item{CntsOfDiffGene}{Threshold for judging the proportion of downregulated genes on the same pathway. For example, when set this value equals to 2, in a particular gene set, there are at least two up-regulated genes more than down regulated genes to judge that there is an up trend for that gene set.}
}
\value{
A BioThemeFinder.ORA_FC object
}
\description{
When gene names and vector measuring the chenge of specific genes, like logFC, are available, create BioThemeFinder object with FC using ORA.
}
\author{
Zhiming Ye
}
