% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{ExtractGenes}
\alias{ExtractGenes}
\title{Extracting genes involved in a specified clustering module}
\usage{
ExtractGenes(x, Cluster)
}
\arguments{
\item{x}{a BioThemeFinder object}

\item{Cluster}{can be MatrixResult or NetworkResult}
}
\value{
A data frame
}
\description{
Extracting genes involved in a specified clustering module
}
\author{
Zhiming Ye.
}
