% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateDupMat.R
\name{GenerateDupMat}
\alias{GenerateDupMat}
\alias{GenerateDupMat,BioThemeFinder.ORA-method}
\alias{GenerateDupMat,BioThemeFinder.ORA_FC-method}
\alias{GenerateDupMat,BioThemeFinder.GSEA-method}
\title{generate repetition rate matrix}
\usage{
GenerateDupMat(x, ...)

\S4method{GenerateDupMat}{BioThemeFinder.ORA}(x, ...)

\S4method{GenerateDupMat}{BioThemeFinder.ORA_FC}(x, ...)

\S4method{GenerateDupMat}{BioThemeFinder.GSEA}(x, ...)
}
\arguments{
\item{x}{BioThemeFinder object}
}
\value{
BioThemeFinder object
}
\description{
generate repetition rate matrix
}
\author{
Zhiming Ye
}
