% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixClustering.R
\name{MatrixClustering}
\alias{MatrixClustering}
\title{Cluster repetition rate matrix based on matrix-based methods}
\usage{
MatrixClustering(
  x,
  k,
  method = "nmf",
  dist_method,
  cluster_method,
  self_def = F
)
}
\arguments{
\item{x}{a BioThemeFinder object}

\item{k}{number of clusters. You can draw PathwayHeatmap to visualize preliminarily and determine the number of clusters}

\item{method}{can be "nmf","hc","pam","fuzzy"}

\item{dist_method}{default:euclidean}

\item{cluster_method}{default:NMF~brunet, hc~ward.D2}

\item{self_def}{whether re-define the dist_method or cluster_method or not}
}
\value{
a BioThemeFinder object
}
\description{
Cluster repetition rate matrix based on matrix-based methods
}
\author{
Zhiming Ye
}
