% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{MultiDBanalysis}
\alias{MultiDBanalysis}
\title{do ORA or GSEA in multi-databases}
\usage{
MultiDBanalysis(
  x,
  PVal = 0.05,
  QVal = 0.1,
  DBlist = c("GO", "KEGG"),
  nGeneCutOff = 3,
  simplify_cutoff = 0.7,
  useMKEGG = T,
  Term2GENE
)
}
\arguments{
\item{x}{a BioThemeFinder object}

\item{PVal}{Pvalue cut-off}

\item{QVal}{Qvalue cut-off (in ORA)}

\item{DBlist}{a vector, Specify in which databases the analysis needs to be completed, can be GO, KEGG, Reactome, SelfDefinedGS, for example :c("GO","KEGG","Reactome")}

\item{nGeneCutOff}{Enriched pathways with a gene counts less than that value were filtered}

\item{simplify_cutoff}{In GO analysis, the cut off of simplify function (from clusterProfiler)}

\item{useMKEGG}{when doing KEGG, if this parameter = T, analysis in both of KEGG pathways and KEGG modules}

\item{Term2GENE}{input self-defined gene collection, one column contains the terms and the other one contains the genes}
}
\value{
A BioThemeFinder object
}
\description{
do ORA or GSEA in multi-databases
}
\author{
Zhiming Ye. This function is powered by clusterProfiler from Guangchuang Yu Lab at southern medical university.
}
