% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkClustering.R
\name{NetworkClustering}
\alias{NetworkClustering}
\title{Cluster repetition rate matrix based on network-based greedy optimization of modularity method}
\usage{
NetworkClustering(x, EdgeCutoff = 0.5, ...)
}
\arguments{
\item{x}{a BioThemeFinder object}

\item{EdgeCutoff}{cutoff of repetition rate}

\item{...}{}
}
\value{
BioThemeFinder object
}
\description{
Cluster repetition rate matrix based on network-based greedy optimization of modularity method
}
\author{
Zhiming Ye
}
