% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteredDotplot.R
\name{PathwayStatsPlot}
\alias{PathwayStatsPlot}
\alias{PathwayStatsPlot,BioThemeFinder.ORA-method}
\alias{PathwayStatsPlot,BioThemeFinder.ORA_FC-method}
\alias{PathwayStatsPlot,BioThemeFinder.GSEA-method}
\title{plot dotplot showing the statstical result of BioThemeFinder}
\usage{
PathwayStatsPlot(
  x,
  clusterType = NULL,
  showStart = NULL,
  showEnd = NULL,
  orderBy,
  ...
)

\S4method{PathwayStatsPlot}{BioThemeFinder.ORA}(
  x,
  clusterType = "NetworkResult",
  showStart = 1,
  showEnd = 20,
  orderBy,
  col_low = "#1e6091",
  col_high = "#99d98c",
  ...
)

\S4method{PathwayStatsPlot}{BioThemeFinder.ORA_FC}(
  x,
  clusterType = "NetworkResult",
  showStart = 1,
  showEnd = 20,
  orderBy,
  col_low = "#1e6091",
  col_high = "#99d98c",
  ...
)

\S4method{PathwayStatsPlot}{BioThemeFinder.GSEA}(
  x,
  clusterType = "NetworkResult",
  showStart = 1,
  showEnd = 20,
  col_low = "#1e6091",
  col_high = "#99d98c",
  ...
)
}
\arguments{
\item{x}{BioThemeFinder object}

\item{clusterType}{can be one of "MatrixResult", "NetworkResult"}

\item{showStart}{showing from <showStart> to <showEnd> in ordered table (ordering is based on the NES)}

\item{showEnd}{showing from <showStart> to <showEnd> in ordered table (ordering is based on the NES)}

\item{orderBy}{can be "pValue","GeneRatio","Counts"}

\item{col_low}{defining color}

\item{col_high}{defining color}
}
\value{
a ggplot2 object, add + facet_grid(~Cluster) to display the clustered result. remove + facet_grid(~othergroup) for unclustered results
}
\description{
plot dotplot showing the statstical result of BioThemeFinder
}
\author{
Zhiming Ye
}
