% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkClustering.R
\name{PlotNetwork}
\alias{PlotNetwork}
\title{plot network}
\usage{
PlotNetwork(x, method = "igraph", Label = T, ...)
}
\arguments{
\item{x}{a BioThemeFinder object}

\item{method}{igraph or ggplot2}

\item{...}{}
}
\description{
plot network
}
\author{
Zhiming Ye
}
