% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praseClusterProfilerRes.R
\name{prase.clusterProfiler.result}
\alias{prase.clusterProfiler.result}
\title{prase result from clusterProfiler}
\usage{
prase.clusterProfiler.result(
  x,
  GOBP = NULL,
  GOCC = NULL,
  GOMF = NULL,
  GOALL = NULL,
  KEGG = NULL,
  MKEGG = NULL,
  Reactome = NULL,
  Self = NULL,
  nGeneCutOff = 3,
  IsSymbol = F,
  OrgDB = NULL
)
}
\arguments{
\item{x}{A BioThemeFinder object}

\item{GOBP}{clusterProfiler object}

\item{GOCC}{clusterProfiler object}

\item{GOMF}{clusterProfiler object}

\item{GOALL}{clusterProfiler object}

\item{KEGG}{clusterProfiler object}

\item{MKEGG}{clusterProfiler object}

\item{Reactome}{clusterProfiler object}

\item{Self}{self-defined gene collection}

\item{nGeneCutOff}{Enriched pathways with a gene counts less than that value were filtered}

\item{IsSymbol}{True~SYMBOL (have done "setreadable" function), not~ENTREZID. NOTE:It is recommended to perform setreadable operation before importing BioThemeFinder}

\item{OrgDB}{the org.DB using to convert ID}
}
\value{
A BioThemeFinder object
}
\description{
prase result from clusterProfiler
}
\author{
Zhiming Ye
}
