% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{resultDF}
\alias{resultDF}
\title{Extracting the enrichment result}
\usage{
resultDF(x, file)
}
\arguments{
\item{x}{a BioThemeFinder object}
}
\value{
A data frame
}
\description{
Extracting the enrichment result
}
\author{
Zhiming Ye
}
